# eth/filters: performance improvement on filterLogs