--- a/eth/filters/filter.go
+++ b/eth/filters/filter.go
@@ -358,6 +358,7 @@ func includes(addresses []common.Address, a common.Address) bool {
 }
 
 // filterLogs creates a slice of logs matching the given criteria.
+// filterLogs expects logs to be in ascending order by block number.
 func filterLogs(logs []*types.Log, fromBlock, toBlock *big.Int, addresses []common.Address, topics [][]common.Hash) []*types.Log {
 	var (
 		ret      []*types.Log