Because the logs have arrived in asc order, so if the `to block` in query criteria is bigger than the current `log.BlockNumber`, then no need to filter the later logs anymore.


Here is a benchmark with:


```bash
cd eth/filters
go test -bench='BenchmarkFilterLogs' . -count 5  -run BenchmarkFilterLogs
```

> the result of master branch


```
goos: darwin
goarch: arm64
pkg: github.com/ethereum/go-ethereum/eth/filters
BenchmarkFilterLogs-8              12616             92477 ns/op
BenchmarkFilterLogs-8              12992             92284 ns/op
BenchmarkFilterLogs-8              12940             91975 ns/op
BenchmarkFilterLogs-8              12987             92502 ns/op
BenchmarkFilterLogs-8              13009             88202 ns/op
```

> the result of this branch

```
goos: darwin
goarch: arm64
pkg: github.com/ethereum/go-ethereum/eth/filters
BenchmarkFilterLogs-8              13528             86444 ns/op
BenchmarkFilterLogs-8              13440             86351 ns/op
BenchmarkFilterLogs-8              14128             86022 ns/op
BenchmarkFilterLogs-8              13968             84665 ns/op
BenchmarkFilterLogs-8              13959             85568 ns/op
```

about 6.2% improvement


BTW,  also add a new function `topicFilter`, used to improve the readability of `filterLogs`