@s1na 2023-08-08
Actually I realized the 6% is improvement on `filterLogs` and not the whole log fetching. IMO not worth changing the code for this. We should explore other avenues for optimizing `getLogs`.